vendor-comparison-engine/requirements.txt
python-dotenv>=1.0.0
pandas>=2.2.0
openpyxl>=3.1.0
rapidfuzz>=3.8.0
PyYAML>=6.0.0
pytest>=7.4.0
streamlit>=1.40.0